Add files via upload
In this project, we used DESeq2 to investigate the differential expression of genes of interests and p-values to see if those differences were significant. Further, a heatmap was used to visualize the DESeq2 data.

Scientific Question: Will hypoxia induce macrophage infiltration in adipose tissue?

Hypothesis: If hypoxia induces obesity-related macrophage infiltration, then we should observe upregulation of CD68 (a marker for macrophage infiltration).